changes to game model and game_routes

+++ app/models/game.py
@@ -7,7 +7,7 @@ class Game(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     title = db.Column(db.String(150), nullable=False, unique=True)
-    price = db.Column(db.Float(4,2), nullable=False)
+    price = db.Column(db.Integer, nullable=False)
     description = db.Column(db.String(2500), nullable=False)
     release_date = db.Column(db.Date, nullable=False)
     is_mature = db.Column(db.Boolean, nullable=False)
